@@ -1,10 +1,14 @@
 import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { importProvidersFrom } from '@angular/core';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NgxsLoggerPluginModule } from '@ngxs/logger-plugin';
+import { NgxsModule } from '@ngxs/store';
 import { setCompodocJson } from '@storybook/addon-docs/angular';
-import { componentWrapperDecorator, moduleMetadata } from '@storybook/angular';
+import { applicationConfig, componentWrapperDecorator } from '@storybook/angular';
 import { MarkdownModule } from 'ngx-markdown';
 
 import { ThemingModule } from '../libs/shared/theming/src';
+import { addState } from './decorators/state';
 import { fixArgTypes } from './fixes/arg-types';
 
 export const parameters = {
@@ -21,17 +25,27 @@ export const parameters = {
 export const argTypesEnhancers = [fixArgTypes()];
 
 export const decorators = [
-  moduleMetadata({
-    imports: [
-      BrowserAnimationsModule,
-      HttpClientModule,
-      MarkdownModule.forRoot({
-        loader: HttpClient,
-      }),
-      ThemingModule,
+  applicationConfig({
+    providers: [
+      importProvidersFrom(
+        BrowserAnimationsModule,
+        HttpClientModule,
+        NgxsModule.forRoot([], { developmentMode: true }),
+        NgxsLoggerPluginModule.forRoot(),
+        MarkdownModule.forRoot({
+          loader: HttpClient,
+        }),
+        ThemingModule
+      ),
     ],
   }),
-  componentWrapperDecorator((story) => `<div class="mat-typography">${story}</div>`),
+  componentWrapperDecorator(
+    (story) => `
+      <div class="mat-typography">${story}</div>
+      <div class="backdrop-filler" style="position: absolute; inset: 0; z-index: -1;"></div>
+    `
+  ),
+  addState(),
 ];
 
 export function setDocs(library: string): void {
