@@ -1,30 +1,38 @@
-import { TestBed } from '@angular/core/testing';
+import { BrowserAnimationsModule, NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { RouterModule } from '@angular/router';
+import { RouterTestingModule } from '@angular/router/testing';
+import { dispatchAction$ } from '@hra-ui/cdk/injectors';
+import { of } from 'rxjs';
+import { Shallow } from 'shallow-render';
 import { AppComponent } from './app.component';
-import { NxWelcomeComponent } from './nx-welcome.component';
+import { initFactory } from './app.init';
+import { AppModule } from './app.module';
+
+jest.mock('@hra-ui/cdk/injectors');
 
 describe('AppComponent', () => {
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [AppComponent, NxWelcomeComponent],
-    }).compileComponents();
+  let shallow: Shallow<AppComponent>;
+
+  beforeEach(() => {
+    shallow = new Shallow(AppComponent, AppModule)
+      .replaceModule(RouterModule, RouterTestingModule)
+      .replaceModule(BrowserAnimationsModule, NoopAnimationsModule);
   });
 
-  it('should create the app', () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.componentInstance;
-    expect(app).toBeTruthy();
+  it('should create component', async () => {
+    expect(shallow.render()).resolves.toBeDefined();
   });
+});
 
-  it(`should have as title 'ftu-ui'`, () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.componentInstance;
-    expect(app.title).toEqual('ftu-ui');
-  });
+describe('initFactory()', () => {
+  const dispatcher = jest.fn().mockReturnValue(of(undefined));
+  jest.mocked(dispatchAction$).mockReturnValue(dispatcher);
 
-  it('should render title', () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const compiled = fixture.nativeElement as HTMLElement;
-    expect(compiled.querySelector('h1')?.textContent).toContain('Welcome ftu-ui');
+  afterEach(() => jest.clearAllMocks());
+
+  it('should dispatch actions', () => {
+    const init = initFactory();
+    init();
+    expect(dispatcher).toHaveBeenCalled();
   });
 });
