@@ -1,2 +1,4 @@
-export { injectOnDestroy } from './lib/on-destroy/on-destroy';
-export { SelectSnapshotSelector, selectQuerySnapshot, selectSnapshot } from './lib/select-snapshot/select-snapshot';
+export { dispatch, dispatch$, dispatchAction, dispatchAction$ } from './lib/dispatch/dispatch';
+export { injectDestroy$ } from './lib/on-destroy/on-destroy';
+export { select$, SelectOptions, StateSelector } from './lib/select/select';
+export { selectQuerySnapshot, selectSnapshot } from './lib/select/select-snapshot';
