@@ -1,3 +1,7 @@
-// Remove once other exports are available
-/** Placeholder to make build and tests pass until any actual library code is created */
-export const placeholder = undefined;
+export * from './lib/biomarker-details/biomarker-details.component';
+export * from './lib/contact-behavior/contact-behavior.component';
+export * from './lib/header-behavior/header-behavior.component';
+export * from './lib/footer-behavior/footer-behavior.component';
+export * from './lib/medical-illustration-behavior/medical-illustration-behavior.component';
+export * from './lib/landing-page-content/landing-page-content.component';
+export * from './lib/tissue-library-behavior/tissue-library-behavior.component';
