@@ -1,7 +1,16 @@
 export * from './lib/biomarker-table-data-card/biomarker-table-data-card.component';
+export * from './lib/biomarker-table-data-icon/biomarker-table-data-icon.component';
+export * from './lib/contact-modal/contact-modal.component';
+export * from './lib/footer/footer.component';
+export * from './lib/fullscreen/fullscreen-container.component';
+export * from './lib/fullscreen/fullscreen-content.component';
+export * from './lib/header/header.component';
 export * from './lib/info-modal/info-modal.component';
+export * from './lib/interactive-svg/interactive-svg.component';
+export * from './lib/landing-page-in-depth/landing-page-in-depth.component';
 export * from './lib/landing-page-intro/landing-page-intro.component';
+export * from './lib/metrics/metrics.component';
 export * from './lib/screen-size-notice/screen-size-notice.component';
 export * from './lib/source-list/source-list.component';
-export * from './lib/metrics/metrics.component';
+export * from './lib/tissue-tree-list/tissue-tree-list.component';
 export * from './lib/title-card/title-card.component';
