@@ -1,6 +1,10 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+import { MatButtonModule } from '@angular/material/button';
+import { MatRippleModule } from '@angular/material/core';
 import { MatIconModule } from '@angular/material/icon';
+import { LinkDirective } from '@hra-ui/cdk';
+import { EMPTY_LINK } from '@hra-ui/cdk';
 
 /**
  * Header component, displays on top of every page.
@@ -8,7 +12,7 @@ import { MatIconModule } from '@angular/material/icon';
 @Component({
   selector: 'hra-header',
   standalone: true,
-  imports: [CommonModule, MatIconModule],
+  imports: [CommonModule, MatButtonModule, MatIconModule, MatRippleModule, LinkDirective],
   templateUrl: './header.component.html',
   styleUrls: ['./header.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
@@ -28,4 +32,10 @@ export class HeaderComponent {
    * Input for app title to show on the left side.
    */
   @Input() appTitle = '';
+
+  /** Input for Product Title link */
+  @Input() productTitleLink = EMPTY_LINK;
+
+  /** Input for about link */
+  @Input() aboutLink = EMPTY_LINK;
 }
