@@ -1,14 +1,15 @@
-import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
-import { CommonModule } from '@angular/common';
+import { EMPTY_LINK, LinkDirective } from '@hra-ui/cdk';
 import { MarkdownModule } from 'ngx-markdown';
 
 /** Component for LandingPage Intro */
 @Component({
   selector: 'hra-landing-page-intro',
   standalone: true,
-  imports: [CommonModule, MatButtonModule, MatIconModule, MarkdownModule],
+  imports: [CommonModule, MatButtonModule, MatIconModule, MarkdownModule, LinkDirective],
   templateUrl: './landing-page-intro.component.html',
   styleUrls: ['./landing-page-intro.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
@@ -29,6 +30,6 @@ export class LandingPageIntroComponent {
   /** Text for more button */
   @Input() moreText = '';
 
-  /** Emits when the more button is clicked */
-  @Output() moreClick = new EventEmitter<void>();
+  /** Link to navigate to on more button click */
+  @Input() moreLink = EMPTY_LINK;
 }
