@@ -1,3 +1,1 @@
-// Remove once other exports are available
-/** Placeholder to make build and tests pass until any actual library code is created */
-export const placeholder = undefined;
+export * from './lib/url/create';
